fix(behavior_path_planner): move findEgoIndex & findEgoSegmentIndex to planner data
